feat: 강의 스킬 가중치 산정 로직을 비율 기반으로 개선 (#75)
* feat: 강의 등록 후 가중치의 기준이 되는 프롬포트 수정, 가중치를 담을 BigDecimal 타입 추가

* refactor: 강의 스킬 가중치 타입 변경, 이에따른 Mapper 파라미터 타입 변경

* feat: 코스 스킬 전체 재집계 대상 조회쿼리 추가

* feat: 코스 스킬 전체 재집계 대상 Mapper 추가

* refactor: 강의 스킬 가중치 정규화 로직 공통화

+++ src/main/java/com/hlinks/domain/course/ai/dto/CourseSummarySkill.java
@@ -3,6 +3,8 @@ package com.hlinks.domain.course.ai.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class CourseSummarySkill {
@@ -10,5 +12,5 @@ public class CourseSummarySkill {
     private String skillName;
     private String sourceSkillName;
     private String newSkillYn;
-    private Integer weight;
+    private BigDecimal weight;
 }
